Refactor audio utilities
- Added embedding_func to litellm_utils for embedding generation.
